mosaic: various fixes around credentials, deleting stuff, etc.

+++ mosaic.py
@@ -6,11 +6,12 @@ from threading import Thread
 from google.auth.transport.requests import AuthorizedSession
 from google.oauth2.credentials import Credentials
 from google_auth_oauthlib.flow import Flow
+from requests import HTTPError
 
 SIZE = 36
 
 
-def get_credentials():
+def get_session():
     flow = Flow.from_client_secrets_file(
         'app-creds.json',
         scopes=['https://www.googleapis.com/auth/photoslibrary.readonly',
@@ -23,10 +24,11 @@ def get_credentials():
     code = input('Enter the authorization code: ')
     token = flow.fetch_token(code=code)
     print(json.dumps(token))
-    print(json.dumps(flow.credentials))
+    # print(json.dumps(flow.credentials))
+    return flow.authorized_session()
 
 
-def list_media_items(session):
+def _actually_list_media_items(session):
     ret = []
     params = {
         'pageSize': 500,
@@ -46,12 +48,6 @@ def list_media_items(session):
             }
         },
     }
-    concurrent = 100
-    download_queue = queue.Queue(concurrent * 2)
-    for i in range(16):
-        t = Thread(target=lambda: _download_image(session, download_queue))
-        t.daemon = True
-        t.start()
 
     while True:
         rsp = session.post(
@@ -63,42 +59,60 @@ def list_media_items(session):
         ret += cur
         print(f'{len(cur)} new items, total {len(ret)}')
 
-        for m in cur:
-            download_queue.put(m)
-
         pageToken = rsp.get('nextPageToken')
         if pageToken is None:
             break
         params['pageToken'] = pageToken
-    download_queue.join()
-
     return ret
 
 
-def _download_image(session, queue):
+def _download_image(session, queue, size):
     while True:
         m = queue.get()
         if m.get('mimeType', '').startswith('image/'):
-            outfile = f"thumbs/{m['id']}.{SIZE}"
+            outfile = f"thumbs/{m['id']}.{size}"
             if not os.path.exists(outfile) or os.stat(outfile).st_size == 0:
-                with open(outfile, 'wb') as out:
-                    r = session.get(m['baseUrl'] + f'=w{SIZE}-h{SIZE}-c')
-                    r.raise_for_status()
-                    out.write(r.content)
+                try:
+                    with open(outfile, 'wb') as out:
+                        r = session.get(m['baseUrl'] + f'=w{size}-h{size}-c')
+                        r.raise_for_status()
+                        out.write(r.content)
+                except HTTPError as e:
+                    print(e)
+                    os.unlink(outfile)
         queue.task_done()
 
 
-def download_media_items(session):
-    media_items = list_media_items(session)
-    with open('media_items.json', 'w') as out:
-        json.dump(media_items, out)
-    print('Wrote', len(media_items), 'media items out')
-    return media_items
+def download_images(session, media_items, size=SIZE):
+    concurrent = 10
+    download_queue = queue.Queue(concurrent * 2)
+    for i in range(16):
+        t = Thread(target=lambda: _download_image(session, download_queue, size))
+        t.daemon = True
+        t.start()
+    used_ids = set()
+    for m in media_items:
+        if m['id'] not in used_ids:
+            download_queue.put(m)
+            used_ids.add(m['id'])
+    download_queue.join()
 
 
-def delete_unknown_files(media_items):
-    known_files = set(f"thumbs/{m['id']}.{SIZE}" for m in media_items)
-    files = set(f'thumbs/{f}' for f in os.listdir('thumbs'))
+def list_media_items(session):
+    if os.path.exists('media_items.json'):
+        with open('media_items.json', 'r') as infile:
+            return json.load(infile)
+    else:
+        media_items = _actually_list_media_items(session)
+        with open('media_items.json', 'w') as out:
+            json.dump(media_items, out)
+        print('Wrote', len(media_items), 'media items out')
+        return media_items
+
+
+def delete_unknown_files(media_items, size=SIZE):
+    known_files = set(f"thumbs/{m['id']}.{size}" for m in media_items)
+    files = set(f'thumbs/{f}' for f in os.listdir('thumbs') if f.endswith(f'.{size}'))
     print(len(files - known_files), 'to delete')
     for f in files - known_files:
         os.unlink(f)
@@ -107,7 +121,9 @@ def delete_unknown_files(media_items):
 if __name__ == "__main__":
     credentials = Credentials.from_authorized_user_file('photos-creds2.json')
     session = AuthorizedSession(credentials)
-    media_items = download_media_items(session)
+    session = get_session()
+    media_items = list_media_items(session)
     print('Read', len(media_items), 'media items')
-
-    delete_unknown_files(media_items)
+    size = 36
+    download_images(session, media_items, size=size)
+    delete_unknown_files(media_items, size=size)
